move the registry helper method into the handlers/registry directory. add consistent tracking information here, so we can track every activity that is fired.
