+++ src/sentry/workflow_engine/handlers/registry/__init__.py
@@ -1,0 +1,5 @@
+from .invoke_activities import invoke_workflow_activity_handlers
+
+__all__ = [
+    "invoke_workflow_activity_handlers",
+]
