+++ tests/sentry/workflow_engine/handlers/workflow/test_workflow_activity_handlers.py
@@ -11,43 +11,14 @@ from sentry.workflow_engine.handlers.workflow.workflow_activity_handlers import 
     seer_activity_handler,
 )
 from sentry.workflow_engine.models import Detector
-from sentry.workflow_engine.registry import (
-    invoke_workflow_activity_handlers,
-    workflow_activity_registry,
-)
+from sentry.workflow_engine.registry import workflow_activity_registry
 from sentry.workflow_engine.typings.grouptype import IssueStreamGroupType
 
 
 class WorkflowActivityRegistryTest(TestCase):
-    def setUp(self) -> None:
-        self.group = self.create_group()
-        self.activity = self.create_group_activity(
-            group=self.group, type=ActivityType.SEER_PR_CREATED.value
-        )
-
     def test_registrants(self) -> None:
         assert "seer_activity" in workflow_activity_registry.registrations
         assert len(workflow_activity_registry.registrations) == 1
-
-    def test_invoke_handlers_safely(self) -> None:
-        handler_a = mock.Mock()
-        handler_b = mock.Mock(side_effect=Exception("Test error"))
-        handler_c = mock.Mock()
-
-        with mock.patch.dict(
-            workflow_activity_registry.registrations,
-            {"handler_a": handler_a, "handler_b": handler_b, "handler_c": handler_c},
-            clear=True,
-        ):
-            invoke_workflow_activity_handlers(self.group, self.activity)
-
-        handler_a.assert_called_once_with(self.group, self.activity, None)
-        handler_b.assert_called_once_with(self.group, self.activity, None)
-        handler_c.assert_called_once_with(self.group, self.activity, None)
-
-    def test_invoke_handlers_no_registrants(self) -> None:
-        with mock.patch.dict(workflow_activity_registry.registrations, {}, clear=True):
-            invoke_workflow_activity_handlers(self.group, self.activity)
 
 
 class SeerActivityHandlerTest(TestCase):
